Add Delete Button
The delete button like the update button will only appear if the
signed in user is the owner of the post. The delete will delete the post
but does not redirect to an updated full list of post as it still
includes the deleted post. Will correct this in next iteration.